requirements.txt
streamlit
streamlit_autorefresh
psycopg2-binary
altair
pandas
numpy
matplotlib
seaborn